Add flag for other plugins to prevent our rendering
It's different than returning 'false' on the event handler, since it
doesn't prevent other plugins from also handling the event. It only
prevents the default rendering performed by gs-noticePreview.

The gs-markdown plugin needs this since it operates on 'rendered'
property of the Notice.

It feels like a bit of a hack tough so we may want to revisit that
either in this plugin or the gs-markdown plugin.